% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{hello}
\alias{hello}
\alias{how_low}
\alias{how_low.character}
\alias{how_low.default}
\title{Hello, world!}
\format{An object of class \code{character} of length 1.}
\usage{
hello

how_low(x)

\\method{how_low}{default}(x)

\\method{how_low}{character}(x)
}
\description{
This is an example function named 'hello'
which prints 'Hello, world!'.
}
\keyword{datasets}

